# imported/hack-govnoprod/infra/openapijson.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Curestry API",
        "description": "AI Prompt Analysis & Optimization Platform",
        "version": "0.1.0"
    },
    "paths": {
        "/analyze/": {
            "post": {
                "tags": [
                    "analysis"
                ],
                "summary": "Analyze Prompt",
                "description": "Analyze a prompt for quality, consistency, and potential improvements.\n\nThis endpoint performs comprehensive prompt analysis including:\n- Language detection and translation\n- Format validation and markup fixing\n- Vocabulary unification\n- Contradiction detection\n- Semantic entropy analysis\n- LLM-as-judge scoring\n- Improvement patch generation\n- Clarification questions",
                "operationId": "analyze_prompt_analyze__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnalyzeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnalyzeResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/analyze/apply": {
            "post": {
                "tags": [
                    "analysis"
                ],
                "summary": "Apply Patches",
                "description": "Apply selected improvement patches to a prompt.",
                "operationId": "apply_patches_analyze_apply_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplyPatchesRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PromptImproved"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/analyze/clarify": {
            "post": {
                "tags": [
                    "analysis"
                ],
                "summary": "Process Clarification",
                "description": "Process clarification answers and provide updated analysis.",
                "operationId": "process_clarification_analyze_clarify_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClarifyRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnalyzeResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AnalyzeRequest": {
                "properties": {
                    "prompt": {
                        "$ref": "#/components/schemas/PromptInput"
                    },
                    "include_entropy": {
                        "type": "boolean",
                        "title": "Include Entropy",
                        "description": "Include semantic entropy analysis",
                        "default": true
                    },
                    "include_clarify": {
                        "type": "boolean",
                        "title": "Include Clarify",
                        "description": "Include clarification questions",
                        "default": true
                    },
                    "include_patches": {
                        "type": "boolean",
                        "title": "Include Patches",
                        "description": "Include improvement patches",
                        "default": true
                    }
                },
                "type": "object",
                "required": [
                    "prompt"
                ],
                "title": "AnalyzeRequest",
                "description": "Request to analyze a prompt."
            },
            "AnalyzeResponse": {
                "properties": {
                    "report": {
                        "$ref": "#/components/schemas/MetricReport"
                    },
                    "patches": {
                        "items": {
                            "$ref": "#/components/schemas/Patch"
                        },
                        "type": "array",
                        "title": "Patches"
                    },
                    "questions": {
                        "items": {
                            "$ref": "#/components/schemas/ClarifyQuestion"
                        },
                        "type": "array",
                        "title": "Questions"
                    }
                },
                "type": "object",
                "required": [
                    "report",
                    "patches",
                    "questions"
                ],
                "title": "AnalyzeResponse",
                "description": "Response from prompt analysis."
            },
            "ApplyPatchesRequest": {
                "properties": {
                    "prompt_id": {
                        "type": "string",
                        "title": "Prompt Id",
                        "description": "ID of the analyzed prompt"
                    },
                    "patch_ids": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Patch Ids",
                        "description": "IDs of patches to apply"
                    },
                    "apply_safe_all": {
                        "type": "boolean",
                        "title": "Apply Safe All",
                        "description": "Apply all safe patches",
                        "default": false
                    }
                },
                "type": "object",
                "required": [
                    "prompt_id",
                    "patch_ids"
                ],
                "title": "ApplyPatchesRequest",
                "description": "Request to apply specific patches."
            },
            "ClarifyAnswer": {
                "properties": {
                    "question_id": {
                        "type": "string",
                        "title": "Question Id",
                        "description": "ID of the question being answered"
                    },
                    "answer": {
                        "type": "string",
                        "title": "Answer",
                        "description": "The answer provided"
                    }
                },
                "type": "object",
                "required": [
                    "question_id",
                    "answer"
                ],
                "title": "ClarifyAnswer",
                "description": "Answer to a clarification question."
            },
            "ClarifyQuestion": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "Question identifier"
                    },
                    "question": {
                        "type": "string",
                        "title": "Question",
                        "description": "The clarification question"
                    },
                    "category": {
                        "type": "string",
                        "title": "Category",
                        "description": "What aspect needs clarification"
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ],
                        "title": "Priority",
                        "description": "Question priority"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "question",
                    "category",
                    "priority"
                ],
                "title": "ClarifyQuestion",
                "description": "Clarification question for ambiguous prompts."
            },
            "ClarifyRequest": {
                "properties": {
                    "prompt_id": {
                        "type": "string",
                        "title": "Prompt Id",
                        "description": "ID of the prompt being clarified"
                    },
                    "answers": {
                        "items": {
                            "$ref": "#/components/schemas/ClarifyAnswer"
                        },
                        "type": "array",
                        "title": "Answers",
                        "description": "Answers to clarification questions"
                    }
                },
                "type": "object",
                "required": [
                    "prompt_id",
                    "answers"
                ],
                "title": "ClarifyRequest",
                "description": "Request to process clarification answers."
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "MetricReport": {
                "properties": {
                    "prompt_id": {
                        "type": "string",
                        "title": "Prompt Id",
                        "description": "Unique identifier for this analysis"
                    },
                    "original_prompt": {
                        "type": "string",
                        "title": "Original Prompt",
                        "description": "The original prompt text"
                    },
                    "analyzed_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Analyzed At"
                    },
                    "detected_language": {
                        "type": "string",
                        "title": "Detected Language",
                        "description": "Auto-detected language"
                    },
                    "translated": {
                        "type": "boolean",
                        "title": "Translated",
                        "description": "Whether prompt was translated",
                        "default": false
                    },
                    "format_valid": {
                        "type": "boolean",
                        "title": "Format Valid",
                        "description": "Whether markup format is valid"
                    },
                    "judge_score": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/MetricScore"
                            }
                        ],
                        "description": "LLM-as-judge evaluation"
                    },
                    "semantic_entropy": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/SemanticEntropy"
                            }
                        ],
                        "description": "Response consistency analysis"
                    },
                    "contradictions": {
                        "items": {
                            "$ref": "#/components/schemas/Contradiction"
                        },
                        "type": "array",
                        "title": "Contradictions"
                    },
                    "length_chars": {
                        "type": "integer",
                        "title": "Length Chars",
                        "description": "Character count"
                    },
                    "length_words": {
                        "type": "integer",
                        "title": "Length Words",
                        "description": "Word count"
                    },
                    "complexity_score": {
                        "type": "number",
                        "maximum": 10.0,
                        "minimum": 0.0,
                        "title": "Complexity Score",
                        "description": "Vocabulary complexity"
                    },
                    "patches": {
                        "items": {
                            "$ref": "#/components/schemas/Patch"
                        },
                        "type": "array",
                        "title": "Patches"
                    },
                    "clarify_questions": {
                        "items": {
                            "$ref": "#/components/schemas/ClarifyQuestion"
                        },
                        "type": "array",
                        "title": "Clarify Questions"
                    },
                    "overall_score": {
                        "type": "number",
                        "maximum": 10.0,
                        "minimum": 0.0,
                        "title": "Overall Score",
                        "description": "Overall quality score"
                    },
                    "improvement_priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ],
                        "title": "Improvement Priority",
                        "description": "How urgently this prompt needs improvement"
                    }
                },
                "type": "object",
                "required": [
                    "prompt_id",
                    "original_prompt",
                    "detected_language",
                    "format_valid",
                    "judge_score",
                    "semantic_entropy",
                    "length_chars",
                    "length_words",
                    "complexity_score",
                    "overall_score",
                    "improvement_priority"
                ],
                "title": "MetricReport",
                "description": "Comprehensive analysis report."
            },
            "MetricScore": {
                "properties": {
                    "score": {
                        "type": "number",
                        "maximum": 10.0,
                        "minimum": 0.0,
                        "title": "Score",
                        "description": "Score from 0-10"
                    },
                    "rationale": {
                        "type": "string",
                        "title": "Rationale",
                        "description": "Brief explanation of the score"
                    },
                    "details": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Details",
                        "description": "Additional metric-specific details"
                    }
                },
                "type": "object",
                "required": [
                    "score",
                    "rationale"
                ],
                "title": "MetricScore",
                "description": "Individual metric score."
            },
            "Patch": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "Unique patch identifier"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "safe",
                            "risky"
                        ],
                        "title": "Type",
                        "description": "Patch risk level"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "markup",
                            "vocabulary",
                            "structure",
                            "clarity"
                        ],
                        "title": "Category",
                        "description": "Type of improvement"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "description": "What this patch does"
                    },
                    "original": {
                        "type": "string",
                        "title": "Original",
                        "description": "Original text to replace"
                    },
                    "improved": {
                        "type": "string",
                        "title": "Improved",
                        "description": "Improved text"
                    },
                    "rationale": {
                        "type": "string",
                        "title": "Rationale",
                        "description": "Why this improvement helps"
                    },
                    "confidence": {
                        "type": "number",
                        "maximum": 1.0,
                        "minimum": 0.0,
                        "title": "Confidence",
                        "description": "Confidence in this patch"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "type",
                    "category",
                    "description",
                    "original",
                    "improved",
                    "rationale",
                    "confidence"
                ],
                "title": "Patch",
                "description": "Improvement suggestion for the prompt."
            },
...
            "PromptImproved": {
                "properties": {
                    "original_prompt": {
                        "type": "string",
                        "title": "Original Prompt",
                        "description": "The original prompt"
                    },
                    "improved_prompt": {
                        "type": "string",
                        "title": "Improved Prompt",
                        "description": "The improved prompt"
                    },
                    "applied_patches": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Applied Patches",
                        "description": "IDs of patches that were applied"
                    },
                    "improvement_summary": {
                        "type": "string",
                        "title": "Improvement Summary",
                        "description": "Summary of changes made"
                    },
                    "quality_gain": {
                        "type": "number",
                        "title": "Quality Gain",
                        "description": "Estimated quality improvement"
                    }
                },
                "type": "object",
                "required": [
                    "original_prompt",
                    "improved_prompt",
                    "applied_patches",
                    "improvement_summary",
                    "quality_gain"
                ],
                "title": "PromptImproved",
                "description": "Final improved prompt after applying patches."
            },
            "PromptInput": {
                "properties": {
                    "content": {
                        "type": "string",
                        "title": "Content",
                        "description": "The prompt text to analyze"
                    },
                    "format_type": {
                        "type": "string",
                        "enum": [
                            "auto",
                            "xml",
                            "markdown"
                        ],
                        "title": "Format Type",
                        "description": "Expected format of the prompt",
                        "default": "auto"
                    },
                    "language": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Language",
                        "description": "Language of the prompt (auto-detected if not provided)"
                    },
                    "metadata": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Metadata",
                        "description": "Additional metadata about the prompt"
                    }
                },
                "type": "object",
                "required": [
                    "content"
                ],
                "title": "PromptInput",
                "description": "Input prompt for analysis."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}